+++ js/const.js
@@ -1,16 +1,11 @@
-export const similarListElement = document.querySelector('.map__canvas');
 export const similarAddTemplate = document.querySelector('#card')
   .content
   .querySelector('.popup');
 export const range = document.querySelector('#range');
 export const form = document.querySelector('.ad-form');
-export const titleFieldset = document.querySelector('#title_fieldset');
 export const roomsField = form.querySelector('[name="rooms"]');
 export const capacityField = form.querySelector('[name="capacity"]');
-export const capacityFieldset = document.querySelector('#capacity_fieldset');
-export const roomsFieldset = document.querySelector('#room_number_fieldset');
 export const priceField = document.querySelector('#price');
-export const priceFieldset = document.querySelector('#price_fieldset');
 export const formInputs = document.querySelectorAll('.ad-form__watch-focus');
 export const timeFieldset = document.querySelector('#time-fieldset');
 export const timeFields = timeFieldset.querySelectorAll('select');
@@ -21,24 +16,11 @@ const filterChildrenSelect = mapFilter.querySelectorAll('select');
 const filterChildrenFieldset = mapFilter.querySelectorAll('fieldset');
 export const allFormAndFilterChildren = [...formChildren, ...filterChildrenSelect, ...filterChildrenFieldset];
 export const submitButton = document.querySelector('.ad-form__submit');
-export const resetButton = document.querySelector('.ad-form__reset');
-export const success = document.querySelector('#success');
-export const body = document.querySelector('.page-body');
-export const errorButton = document.querySelector('.error__button');
-export const error = document.querySelector('.error');
 export const isEscapeKey = (evt) => evt.key === 'Escape';
 
 export const valuesDivs = [
   document.querySelector('#price')
 ];
-
-export const types = {
-  flat: 'Квартира',
-  bungalow: 'Бунгало',
-  house: 'Дом',
-  palace: 'Дворец',
-  hotel: 'Отель'
-};
 
 export const roomsCount = {
   'alone': 'комната',
@@ -63,12 +45,4 @@ export const guestsOption = {
   '2': ['2', '3'],
   '3': ['3'],
   '0': ['100']
-};
-
-export const typesMinPrice = {
-  bungalow: '0',
-  flat: '1000',
-  hotel: '3000',
-  house: '5000',
-  palace: '10000'
 };
