+++ js/main.js
@@ -4,7 +4,7 @@ import './validate-form.js';
 import {interactionWithForm} from './form.js';
 import {getData} from './api.js';
 import {onMainPinMoveend} from './utils.js';
-import {placeAddsLabels} from './fill-map.js';
+import {placeAddsLabels} from './map.js';
 import {showDataError} from './show-error-or-success.js';
 
 // Деактивировали интерфейс
