+++ js/map.js
@@ -1,6 +1,7 @@
-import {pinIcon, map} from './libs/leaflet-init.js';
-import {SIMILAR_ADD_COUNT} from './preset-const.js';
+import {pinIcon, map, mainPin} from './libs/leaflet-init.js';
+import {SIMILAR_ADD_COUNT, BASE_COORDS} from './preset-const.js';
 import {setAdd} from './set-similar-adds.js';
+import {setAddressInput} from './utils.js';
 
 // Размещение меток объявлений на карте
 export const placeAddsLabels = (array) => {
@@ -30,4 +31,19 @@ export const placeAddsLabels = (array) => {
     });
 };
 
+// Сбрасываем маркер в начальное состояние
+export const resetPin = () => {
+  mainPin.setLatLng(
+    BASE_COORDS
+  );
+};
 
+export const onMainPinMoveend = (evt) => {
+
+  const newCoords = evt.target.getLatLng();
+
+  const lat = newCoords.lat.toFixed(5);
+  const lng = newCoords.lng.toFixed(5);
+
+  setAddressInput(lat, lng);
+};
