+++ js/preset-const.js
@@ -1,67 +1,27 @@
-export const TITLES = [
-  'Апартаменты-студия, 19 м²',
-  'Квартира, 111 м²',
-  'Квартира, 350 м²'
-];
-
-export const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel'
-];
-
-export const CHECKIN = [
-  '12:00',
-  '13:00',
-  '14:00'
-];
-
-export const CHECKOUT = [
-  '12:00',
-  '13:00',
-  '14:00'
-];
-
-export const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner'
-];
-
-export const DESCRIPTIONS = [
-  'Рядом большое количество магазинов',
-  'До метро 1 мин пешком, или 2 остановки на трамвае, остановка у дома',
-  'Проблем с парковкой нет, идеальная транспортная доступность, до 3го кольца 10 мин на машине'
-];
-
-export const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
-];
-
-export const AVATAR_ARRAY = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
-
 export const SIMILAR_ADD_COUNT = 10;
 export const MAX_PRICE = 10000;
-export const MAX_ROOMS_COUNT = 10;
-export const MAX_GUESTS_COUNT = 10;
-export const MIN_LAT_POINT = 35.65000;
-export const MAX_LAT_POINT = 35.70000;
-export const MIN_LNG_POINT = 139.70000;
-export const MAX_LNG_POINT = 139.80000;
 export const COORDINATE_FRACTIONAL_DIGITS = 5;
 export const COUNT_FOR_CORRECT_END = 4;
-export const AVATAR_PATH = 'img/avatars/user';
 
 export const BASE_COORDS = {
   lat: 35.67500,
   lng: 139.75000
+};
+
+export const TYPES = {
+  flat: 'Квартира',
+  bungalow: 'Бунгало',
+  house: 'Дом',
+  palace: 'Дворец',
+  hotel: 'Отель'
+};
+
+export const TYPES_NIN_PRICE = {
+  bungalow: '0',
+  flat: '1000',
+  hotel: '3000',
+  house: '5000',
+  palace: '10000'
 };
 
 export const ZOOM = 12;
