+++ js/utils.js
@@ -1,74 +1,5 @@
-import {timeFields, roomsCount, guestsCount, submitButton} from './const.js';
-import {AVATAR_ARRAY, AVATAR_PATH, COUNT_FOR_CORRECT_END, BASE_COORDS} from './preset-const.js';
-import {pristine} from './libs/pristin-init.js';
-import {mainPin} from './libs/leaflet-init.js';
-
-// Случайное целое положительное число
-export const getRandomWholeNumber = (startNumber, endNumber) => {
-
-  if (startNumber < 0 || endNumber < 0) {
-    return false;
-  }
-
-  if (startNumber === endNumber) {
-    return startNumber;
-  }
-
-  const maxWholeNumber = Math.floor(Math.max(startNumber, endNumber));
-  const minWholeNumber = Math.ceil(Math.min(startNumber, endNumber));
-  const randomWholeNumber = Math.floor(Math.random() * (maxWholeNumber - minWholeNumber + 1)) + minWholeNumber;
-
-  return randomWholeNumber;
-
-};
-
-// Случайное целое положительное число с плавающей точкой
-export const getRandomFractionalNumber = (startNumber, endNumber, fractionalDigits = 2) => {
-
-  if (startNumber < 0 || endNumber < 0) {
-    return false;
-  }
-
-  if (startNumber === endNumber) {
-    return startNumber;
-  }
-
-  const maxNumber = Math.max(startNumber, endNumber);
-  const minNumber = Math.min(startNumber, endNumber);
-  const randomNumber = Math.random() * (maxNumber - minNumber) + minNumber;
-
-  return randomNumber.toFixed(fractionalDigits);
-
-};
-
-// Случайный элемент массива, может повторяться
-export const getRandomElementFromArray = (elements) => elements[getRandomWholeNumber(0, elements.length - 1)];
-
-// Собрать случайный массив из элементов другого массива (может быть пустым)
-export const getRandomArrayWithNull = (array) => {
-
-  const randomArray = [];
-  const maxNumber = getRandomWholeNumber(0, array.length - 1) - 1;
-
-  for (let j = 0; j <= maxNumber; j++) {
-    randomArray.push(array[j]);
-  }
-
-  return randomArray;
-};
-
-// Собрать случайный массив из элементов другого массива (не может быть пустым)
-export const getRandomArray = (array) => {
-
-  const randomArray = [];
-  const maxNumber = getRandomWholeNumber(0, array.length - 1);
-
-  for (let j = 0; j <= maxNumber; j++) {
-    randomArray.push(array[j]);
-  }
-
-  return randomArray;
-};
+import {roomsCount, guestsCount} from './const.js';
+import {COUNT_FOR_CORRECT_END} from './preset-const.js';
 
 // Добавляем фотографии
 export const addPhotos = (array, blockPhoto, blockPhotos) => {
@@ -96,22 +27,6 @@ export const getCorrectKeyForGuests = (count) => {
   } return guestsCount.many;
 };
 
-// Получить массив из путей до картинок
-export const arrayAvatarPath = [];
-
-export const getOrderedPictures = (array, path) => {
-  array.forEach((item) => {
-    if (item >= 10) {
-      arrayAvatarPath.push(`${path}${item}.png`);
-    } else {
-      arrayAvatarPath.push(`${path}0${item}.png`);
-    }
-  });
-  return arrayAvatarPath;
-};
-
-getOrderedPictures(AVATAR_ARRAY, AVATAR_PATH);
-
 // Подставляем текстовый контент
 export const setTextContent = (obj) => {
   Object.entries(obj).forEach(([, value]) => {
@@ -128,57 +43,9 @@ export const removeBlock = (obj) => {
   });
 };
 
-// Функция для удаления ошибок пристины
-export const deleteErrors = (evt) => {
-
-  const errorMessage = evt.target.nextElementSibling;
-
-  if (errorMessage.matches('.pristine-error')) {
-
-    pristine.reset();
-
-  }
-};
-
-// Синхронизация значений селектов
-export function setValue (evt) {
-
-  timeFields.forEach((timeField) => {
-    timeField.value = evt.target.value;
-  });
-
-}
-
-// Сбрасываем маркер в начальное состояние
-export const resetPin = () => {
-  mainPin.setLatLng(
-    BASE_COORDS
-  );
-};
-
 // Добавляет координаты в поле адреса
-const setAddressInput = (lat,lng) => {
+export const setAddressInput = (lat,lng) => {
   const addressInput = document.querySelector('#address');
 
   addressInput.value = `${lat}, ${lng}`;
-};
-
-export const blockSubmitButton = () => {
-  submitButton.disabled = true;
-  submitButton.textContent = 'Отправляю';
-};
-
-export const unblockSubmitButton = () => {
-  submitButton.disabled = false;
-  submitButton.textContent = 'Опубликовать';
-};
-
-export const onMainPinMoveend = (evt) => {
-
-  const newCoords = evt.target.getLatLng();
-
-  const lat = newCoords.lat.toFixed(5);
-  const lng = newCoords.lng.toFixed(5);
-
-  setAddressInput(lat, lng);
 };
